scripts/requirements.txt
finance-datareader
pandas
supabase
python-dotenv
numpy
pykrx
requests
selenium
webdriver-manager
beautifulsoup4
lxml
setuptools